arreglando el migration que cambia articulo_combo para articulo_combo_id para que solo lo haga si no existe

+++ db/migrate/20230718201438_change_articulo_combo_to_reference.rb
@@ -3,17 +3,21 @@ class ChangeArticuloComboToReference < ActiveRecord::Migration[7.0]
 	def change
 		reversible do |dir|
 			dir.up do
-				add_column :formulas_productos_terminados, :articulo_combo_id, :bigint
-				FormulasProductosTerminado.reset_column_information
-				add_foreign_key :formulas_productos_terminados, :articulos, column: :articulo_combo_id, foreign_key: { to_table: :articulos }
+        if !column_exists?(:formulas_productos_terminados, :articulo_combo_id) || !foreign_key_exists?(:formulas_productos_terminados, :articulo_combo_id)
+          add_column :formulas_productos_terminados, :articulo_combo_id, :bigint unless column_exists?(:formulas_productos_terminados, :articulo_combo_id)
+          FormulasProductosTerminado.reset_column_information
 
-				comboMappingUp = FormulasProductosTerminado.pluck(:id, :articulo_combo).to_h
 
-				FormulasProductosTerminado.find_each do |fpt|
-					fpt.update_attribute(:articulo_combo_id, comboMappingUp[fpt.id])
-				end
+          add_foreign_key :formulas_productos_terminados, :articulos, column: :articulo_combo_id, foreign_key: { to_table: :articulos } unless foreign_key_exists?(:formulas_productos_terminados, :articulo_combo_id)
 
-				remove_column :formulas_productos_terminados, :articulo_combo, :integer
+          comboMappingUp = FormulasProductosTerminado.pluck(:id, :articulo_combo).to_h
+
+          FormulasProductosTerminado.find_each do |fpt|
+            fpt.update_attribute(:articulo_combo_id, comboMappingUp[fpt.id])
+          end
+
+          remove_column :formulas_productos_terminados, :articulo_combo, :integer
+			  end
 			end
 
 			dir.down do
